+++ packages/atomic/src/components/result-lists/atomic-folded-result-list/atomic-folded-result-list.tsx
@@ -22,11 +22,13 @@ import {
 } from '../../atomic-result/atomic-result-display-options';
 import {ResultListCommon, ResultRenderingFunction} from '../result-list-common';
 import {FoldedResultListStateContextEvent} from '../result-list-decorators';
+import {randomID} from '../../../utils/utils';
 
 /**
  * The `atomic-folded-result-list` component is responsible for displaying folded query results, by applying one or more result templates for up to three layers (i.e., to the result, child and grandchild).
  *
  * @part result-list - The element containing every result of a result list
+ * @part outline - The element displaying an outline or a divider around a result
  */
 @Component({
   tag: 'atomic-folded-result-list',
@@ -56,6 +58,7 @@ export class AtomicFoldedResultList implements InitializableComponent {
 
   public resultListCommon!: ResultListCommon;
   private renderingFunction: ((res: FoldedResult) => HTMLElement) | null = null;
+  private loadingFlag = randomID('firstResultLoaded-');
 
   /**
    * Whether to automatically retrieve an additional page of results and append it to the
@@ -146,6 +149,7 @@ export class AtomicFoldedResultList implements InitializableComponent {
       onError: () => {
         this.templateHasError = true;
       },
+      loadingFlag: this.loadingFlag,
     });
 
     this.foldedResultList = this.initFolding(
@@ -173,13 +177,6 @@ export class AtomicFoldedResultList implements InitializableComponent {
     foldedResult: FoldedResult
   ): HTMLElement | DocumentFragment {
     return this.resultListCommon.getContentOfResultTemplate(foldedResult);
-  }
-
-  public componentDidRender() {
-    this.resultListCommon.componentDidRender(
-      this.foldedResultListState.firstSearchExecuted,
-      this.listWrapperRef
-    );
   }
 
   public render() {