+++ packages/atomic/src/components/search-box-suggestions/atomic-search-box-recent-queries/atomic-search-box-recent-queries.tsx
@@ -7,7 +7,9 @@ import {
 import {Component, Element, Prop, State, h} from '@stencil/core';
 import {
   dispatchSearchBoxSuggestionsEvent,
+  SearchBoxDividerElement,
   SearchBoxSuggestionElement,
+  SearchBoxSuggestionItem,
   SearchBoxSuggestionsBindings,
 } from '../suggestions-common';
 import {once} from '../../../utils/utils';
@@ -15,6 +17,9 @@ import {SafeStorage, StorageItems} from '../../../utils/local-storage-utils';
 
 /**
  * The `atomic-search-box-recent-queries` component can be added as a child of an `atomic-search-box` component, allowing for the configuration of recent query suggestions.
+ *
+ * @part recent-query-title - The 'Recent queries' title.
+ * @part recent-query-clear - The 'Clear' button for clearing recent queries    .
  */
 @Component({
   tag: 'atomic-search-box-recent-queries',
@@ -28,6 +33,15 @@ export class AtomicSearchBoxRecentQueries {
   @Element() private host!: HTMLElement;
 
   @State() public error!: Error;
+
+  /**
+   * The SVG icon to display.
+   *
+   * - Use a value that starts with `http://`, `https://`, `./`, or `../`, to fetch and display an icon from a given location.
+   * - Use a value that starts with `assets://`, to display an icon from the Atomic package.
+   * - Use a stringified SVG to display it directly.
+   */
+  @Prop() public icon?: string;
 
   /**
    * The maximum number of suggestions that will be displayed if the user has typed something into the input field.
@@ -47,6 +61,10 @@ export class AtomicSearchBoxRecentQueries {
     } catch (error) {
       this.error = error as Error;
     }
+  }
+
+  private renderIcon() {
+    return this.icon || Clock;
   }
 
   private initialize() {
@@ -91,7 +109,7 @@ export class AtomicSearchBoxRecentQueries {
     this.storage.removeItem(StorageItems.RECENT_QUERIES);
   }
 
-  private renderItems(): SearchBoxSuggestionElement[] {
+  private renderItems(): SearchBoxSuggestionItem[] {
     if (!this.recentQueriesList.state.analyticsEnabled) {
       return [];
     }
@@ -107,8 +125,8 @@ export class AtomicSearchBoxRecentQueries {
       )
       .slice(0, max);
 
-    const suggestionElements = filteredQueries.map((value) =>
-      this.renderItem(value)
+    const suggestionElements: SearchBoxSuggestionItem[] = filteredQueries.map(
+      (value) => this.renderItem(value)
     );
     if (suggestionElements.length) {
       suggestionElements.unshift(this.renderClear());
@@ -117,15 +135,15 @@ export class AtomicSearchBoxRecentQueries {
     return suggestionElements;
   }
 
-  private renderClear(): SearchBoxSuggestionElement {
+  private renderClear(): SearchBoxDividerElement {
     return {
       key: 'recent-query-clear',
       content: (
         <div class="flex justify-between w-full">
-          <span class="font-bold">
+          <span class="font-bold" part="recent-query-title">
             {this.bindings.i18n.t('recent-searches')}
           </span>
-          <span>{this.bindings.i18n.t('clear')}</span>
+          <span part="recent-query-clear">{this.bindings.i18n.t('clear')}</span>
         </div>
       ),
       onSelect: () => {
@@ -142,7 +160,10 @@ export class AtomicSearchBoxRecentQueries {
       query: value,
       content: (
         <div class="flex items-center break-all">
-          <atomic-icon icon={Clock} class="w-4 h-4 mr-2 shrink-0"></atomic-icon>
+          <atomic-icon
+            icon={this.renderIcon()}
+            class="w-4 h-4 mr-2 shrink-0"
+          ></atomic-icon>
           {query === '' ? (
             <span class="break-all line-clamp-2">{value}</span>
           ) : (