+++ packages/atomic/src/utils/store.ts
@@ -32,6 +32,13 @@ export type AtomicStore = {
   facetElements: HTMLElement[];
   sortOptions: SortDropdownOption[];
   iconAssetsPath: string;
+  /**
+   * Dynamic list of flags that allows for syncing various components loading state on initialization.
+   * E.g., waiting for result template component to be rendered.
+   *
+   * This is a better indicator than only the "firstSearchExecuted" of the Headless search state.
+   */
+  loadingFlags: string[];
 };
 
 export const initialStore: () => AtomicStore = () => ({
@@ -42,6 +49,7 @@ export const initialStore: () => AtomicStore = () => ({
   facetElements: [],
   sortOptions: [],
   iconAssetsPath: '',
+  loadingFlags: [],
 });
 
 export const registerFacetToStore = <T extends FacetType, U extends string>(
@@ -82,3 +90,41 @@ export const getAllFacets = (store: ObservableMap<AtomicStore>) => ({
   ...store.state.categoryFacets,
   ...store.state.numericFacets,
 });
+
+export const setLoadingFlag = (
+  store: ObservableMap<AtomicStore>,
+  flag: string
+) => {
+  const flags = store.get('loadingFlags');
+  store.set('loadingFlags', flags.concat(flag));
+};
+
+export const unsetLoadingFlag = (
+  store: ObservableMap<AtomicStore>,
+  flag: string
+) => {
+  const flags = store.get('loadingFlags');
+  store.set(
+    'loadingFlags',
+    flags.filter((value) => value !== flag)
+  );
+};
+
+export const hasLoadingFlag = (
+  store: ObservableMap<AtomicStore>,
+  flag: string
+) => store.get('loadingFlags').indexOf(flag) !== -1;
+
+export const waitUntilAppLoaded = (
+  store: ObservableMap<AtomicStore>,
+  callback: () => void
+) => {
+  store.onChange('loadingFlags', (flags) => {
+    if (!flags.length) {
+      callback();
+    }
+  });
+};
+
+export const isAppLoaded = (store: ObservableMap<AtomicStore>) =>
+  !store.get('loadingFlags').length;