+++ packages/headless/src/controllers/recommendation/context/headless-recommendation-context.test.ts
@@ -1,0 +1,19 @@
+import {buildContext, Context} from './headless-recommendation-context';
+import {
+  buildMockRecommendationAppEngine,
+  MockRecommendationEngine,
+} from '../../../test/mock-engine';
+
+describe('Context', () => {
+  let context: Context;
+  let engine: MockRecommendationEngine;
+
+  beforeEach(() => {
+    engine = buildMockRecommendationAppEngine();
+    context = buildContext(engine);
+  });
+
+  it('initializes properly', () => {
+    expect(context.state.values).toEqual({});
+  });
+});