+++ packages/samples/atomic-react/src/pages/FoldedResultListWithCustomChildrenPage.tsx
@@ -81,30 +81,32 @@ function MyTemplate(result: FoldedResult) {
       </AtomicResultSectionChildren>
       <AtomicResultSectionBottomMetadata>
         <AtomicResultFieldsList>
-          {result.result.raw.cat_platform && (
-            <>
-              <span className="field-label">
-                <AtomicText value="Platform" />
-              </span>
-              <AtomicResultText field="cat_platform" />
-            </>
-          )}
-          {result.result.raw.cat_condition && (
-            <>
-              <span className="field-label">
-                <AtomicText value="Condition" />
-              </span>
-              <AtomicResultText field="cat_condition" />
-            </>
-          )}
-          {result.result.raw.cat_categories && (
-            <>
-              <span className="field-label">
-                <AtomicText value="Tags" />
-              </span>
-              <AtomicResultMultiValueText field="cat_categories" />
-            </>
-          )}
+          <>
+            {result.result.raw.cat_platform && (
+              <>
+                <span className="field-label">
+                  <AtomicText value="Platform" />
+                </span>
+                <AtomicResultText field="cat_platform" />
+              </>
+            )}
+            {result.result.raw.cat_condition && (
+              <>
+                <span className="field-label">
+                  <AtomicText value="Condition" />
+                </span>
+                <AtomicResultText field="cat_condition" />
+              </>
+            )}
+            {result.result.raw.cat_categories && (
+              <>
+                <span className="field-label">
+                  <AtomicText value="Tags" />
+                </span>
+                <AtomicResultMultiValueText field="cat_categories" />
+              </>
+            )}
+          </>
         </AtomicResultFieldsList>
       </AtomicResultSectionBottomMetadata>
     </>