+++ packages/atomic/CHANGELOG.md
@@ -3,6 +3,102 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.61.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.60.1...@coveo/atomic@1.61.0) (2022-05-19)
+
+
+### Bug Fixes
+
+* **atomic:** fix grid columns when image is not large ([#2000](https://github.com/coveo/ui-kit/issues/2000)) ([a73fa8a](https://github.com/coveo/ui-kit/commit/a73fa8a0276a82223792d5a4c13485adfaecff32))
+* **atomic:** fix regression on deprecated image prop for result list ([#2001](https://github.com/coveo/ui-kit/issues/2001)) ([c632c69](https://github.com/coveo/ui-kit/commit/c632c69480468b3cef144cd099a95819f69e0336))
+
+
+### Features
+
+* **atomic:** made result list outline and divider easily customizable ([#1998](https://github.com/coveo/ui-kit/issues/1998)) ([f2c1d1f](https://github.com/coveo/ui-kit/commit/f2c1d1fbfe0ad83b324fbe5567909b419df84ab8))
+
+
+
+
+
+## [1.60.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.60.0...@coveo/atomic@1.60.1) (2022-05-18)
+
+**Note:** Version bump only for package @coveo/atomic
+
+
+
+
+
+# [1.60.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.59.1...@coveo/atomic@1.60.0) (2022-05-18)
+
+
+### Features
+
+* **atomic:** added analytics & highlighting to smart snippet suggestions ([#1990](https://github.com/coveo/ui-kit/issues/1990)) ([7da3924](https://github.com/coveo/ui-kit/commit/7da3924b8effa71dc858f5968f8c161aeda16dd2))
+
+
+
+
+
+## [1.59.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.59.0...@coveo/atomic@1.59.1) (2022-05-18)
+
+
+### Bug Fixes
+
+* **atomic:** clone child nodes for inherited templates in children ([#1995](https://github.com/coveo/ui-kit/issues/1995)) ([25097eb](https://github.com/coveo/ui-kit/commit/25097eb7c0181a7da3bbcc9e53da4d1603012610))
+
+
+
+
+
+# [1.59.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.58.0...@coveo/atomic@1.59.0) (2022-05-17)
+
+
+### Features
+
+* **atomic:** allow to customise suggestions icons ([#1980](https://github.com/coveo/ui-kit/issues/1980)) ([b5f7a6d](https://github.com/coveo/ui-kit/commit/b5f7a6de0315136551178b7a8fa69c92920a49dc))
+
+
+
+
+
+# [1.58.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.57.0...@coveo/atomic@1.58.0) (2022-05-16)
+
+
+### Features
+
+* **atomic:** initial load performance + UX improvements ([#1984](https://github.com/coveo/ui-kit/issues/1984)) ([8f95c38](https://github.com/coveo/ui-kit/commit/8f95c385ef1b834f2befa63d7be2e34b977dac85))
+
+
+
+
+
+# [1.57.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.56.1...@coveo/atomic@1.57.0) (2022-05-16)
+
+
+### Bug Fixes
+
+* **atomic:** small issues related to folding ([#1986](https://github.com/coveo/ui-kit/issues/1986)) ([5f8567b](https://github.com/coveo/ui-kit/commit/5f8567bdc99c9ce8336b5f03b8c95aef1ce2adb3))
+
+
+### Features
+
+* **atomic:** add parts to suggestions ([#1982](https://github.com/coveo/ui-kit/issues/1982)) ([f4d4463](https://github.com/coveo/ui-kit/commit/f4d446344dbf167b61737cda98c6ebc6ea262ee0))
+
+
+
+
+
+## [1.56.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.56.0...@coveo/atomic@1.56.1) (2022-05-13)
+
+
+### Bug Fixes
+
+* **atomic:** highlight encoding issue w/ result-text & no-results ([#1987](https://github.com/coveo/ui-kit/issues/1987)) ([4b511be](https://github.com/coveo/ui-kit/commit/4b511be6b1e24ec950c6e112bc005d82a7bec223))
+
+
+
+
+
 # [1.56.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.55.1...@coveo/atomic@1.56.0) (2022-05-12)