+++ packages/samples/headless-react/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.52.7](https://github.com/coveo/ui-kit/compare/headless-react@0.52.6...headless-react@0.52.7) (2022-05-19)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+## [0.52.6](https://github.com/coveo/ui-kit/compare/headless-react@0.52.5...headless-react@0.52.6) (2022-05-18)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+## [0.52.5](https://github.com/coveo/ui-kit/compare/headless-react@0.52.4...headless-react@0.52.5) (2022-05-17)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+## [0.52.4](https://github.com/coveo/ui-kit/compare/headless-react@0.52.3...headless-react@0.52.4) (2022-05-13)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
 ## [0.52.3](https://github.com/coveo/ui-kit/compare/headless-react@0.52.2...headless-react@0.52.3) (2022-05-11)
 
 **Note:** Version bump only for package headless-react